+++ packages/enzyme-mocha/.npmignore
@@ -1,0 +1,1 @@
+# Left empty in order to include the build directory for npm publish
