+++ packages/enzyme-mocha/test/jsdom-spec.js
@@ -1,0 +1,3 @@
+describe('jsdom', () => {
+  it.skip('TODO: set up tests');
+});
