+++ scripts/test-packages.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env zsh
+
+# die on error
+setopt -e
+
+for pkg in packages/*/ ; do
+  pushd ${pkg} > /dev/null
+  npm test
+  popd > /dev/null
+done
