+++ src/index.js
@@ -1,71 +1,7 @@
 import cheerio from 'cheerio';
-import Sinon from 'sinon';
 import ReactWrapper from './ReactWrapper';
 import ShallowWrapper from './ShallowWrapper';
-import { onPrototype } from './Utils';
 import { renderToStaticMarkup } from './react-compat';
-
-/**
- * @class Enzyme
- */
-
-let jsdom;
-try {
-  require('jsdom'); // could throw
-  jsdom = require('mocha-jsdom');
-} catch (e) {
-  // jsdom is not supported...
-}
-
-export let sinon = Sinon.sandbox.create();
-
-export function describeWithDOM(a, b) {
-  describe('(uses jsdom)', () => {
-    if (typeof jsdom === 'function') {
-      jsdom();
-      describe(a, b);
-    } else {
-      // if jsdom isn't available, skip every test in this describe context
-      describe.skip(a, b);
-    }
-  });
-}
-
-export function useSetStateHack() {
-  let cleanup = false;
-  before(() => {
-    if (typeof global.document === 'undefined') {
-      cleanup = true;
-      global.document = {};
-    }
-  });
-  after(() => {
-    if (cleanup) {
-      delete global.document;
-    }
-  });
-}
-
-export function spySetup() {
-  sinon = Sinon.sandbox.create();
-}
-
-export function spyTearDown() {
-  sinon.restore();
-}
-
-export function useSinon() {
-  beforeEach(spySetup);
-  afterEach(spyTearDown);
-}
-
-export function spyLifecycle(Component) {
-  onPrototype(Component, (proto, name) => sinon.spy(proto, name));
-}
-
-export function spyMethods(Component) {
-  onPrototype(Component, null, (proto, name) => sinon.spy(proto, name));
-}
 
 /**
  * Mounts and renders a react component into the document and provides a testing wrapper around it.
