+++ test/ReactWrapper-spec.js
@@ -1,11 +1,15 @@
 import React from 'react';
 import { expect } from 'chai';
-import { shallow, render, ShallowWrapper } from '../';
+import {
+  mount,
+  render,
+  ReactWrapper,
+} from '../src/';
 import sinon from 'sinon';
-import { describeIf } from './_helpers';
-import { REACT013 } from '../version';
+import { describeIf, describeWithDOM } from './_helpers';
+import { REACT013 } from '../src/version';
 
-describe('shallow', () => {
+describeWithDOM('mount', () => {
 
   describeIf(!REACT013, 'stateless components', () => {
     it('works with stateless components', () => {
@@ -15,8 +19,8 @@ describe('shallow', () => {
           <div className="qoo">{foo}</div>
         </div>
       );
-      const wrapper = shallow(<Foo foo="qux" />);
-      expect(wrapper.type()).to.equal('div');
+      const wrapper = mount(<Foo foo="qux" />);
+      expect(wrapper.type()).to.equal(Foo);
       expect(wrapper.find('.bar')).to.have.length(1);
       expect(wrapper.find('.qoo').text()).to.equal('qux');
     });
@@ -28,12 +32,12 @@ describe('shallow', () => {
       const a = <div className="foo" />;
       const b = <div className="foo" />;
       const c = <div className="bar" />;
-      expect(shallow(a).contains(b)).to.be.true;
-      expect(shallow(a).contains(c)).to.be.false;
+      expect(mount(a).contains(b)).to.be.true;
+      expect(mount(a).contains(c)).to.be.false;
     });
 
     it('should allow matches on a nested node', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo" />
         </div>
@@ -46,7 +50,7 @@ describe('shallow', () => {
       class Foo extends React.Component {
         render() { return <div />; }
       }
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <Foo />
         </div>
@@ -60,7 +64,7 @@ describe('shallow', () => {
   describe('.find(selector)', () => {
 
     it('should find an element based on a class name', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <input className="foo" />
         </div>
@@ -69,7 +73,7 @@ describe('shallow', () => {
     });
 
     it('should find an element based on a tag name', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <input className="foo" />
         </div>
@@ -81,7 +85,7 @@ describe('shallow', () => {
       class Foo extends React.Component {
         render() { return <div />; }
       }
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <Foo className="foo" />
         </div>
@@ -90,7 +94,7 @@ describe('shallow', () => {
     });
 
     it('should find multiple elements based on a class name', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <input className="foo" />
           <button className="foo" />
@@ -100,7 +104,7 @@ describe('shallow', () => {
     });
 
     it('should find multiple elements based on a tag name', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <input className="foo" />
           <input />
@@ -112,7 +116,7 @@ describe('shallow', () => {
     });
 
     it('should find multiple elements based on a constructor', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <input className="foo" />
           <input />
@@ -124,7 +128,7 @@ describe('shallow', () => {
     });
 
     it('should throw on a complex selector', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <input className="foo" />
           <input />
@@ -139,7 +143,7 @@ describe('shallow', () => {
   describe('.findWhere(predicate)', () => {
 
     it('should return all elements for a truthy test', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <input className="foo" />
           <input />
@@ -149,7 +153,7 @@ describe('shallow', () => {
     });
 
     it('should return no elements for a falsy test', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <input className="foo" />
           <input />
@@ -159,7 +163,7 @@ describe('shallow', () => {
     });
 
     it('should call the predicate with the wrapped node as the first argument', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo bar" />
           <div className="foo baz" />
@@ -172,10 +176,10 @@ describe('shallow', () => {
       const spy = sinon.spy(stub);
       wrapper.findWhere(spy);
       expect(spy.callCount).to.equal(4);
-      expect(spy.args[0][0]).to.be.instanceOf(ShallowWrapper);
-      expect(spy.args[1][0]).to.be.instanceOf(ShallowWrapper);
-      expect(spy.args[2][0]).to.be.instanceOf(ShallowWrapper);
-      expect(spy.args[3][0]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[0][0]).to.be.instanceOf(ReactWrapper);
+      expect(spy.args[1][0]).to.be.instanceOf(ReactWrapper);
+      expect(spy.args[2][0]).to.be.instanceOf(ReactWrapper);
+      expect(spy.args[3][0]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[1][0].hasClass('bar')).to.be.true;
       expect(spy.args[2][0].hasClass('baz')).to.be.true;
       expect(spy.args[3][0].hasClass('bux')).to.be.true;
@@ -195,9 +199,9 @@ describe('shallow', () => {
           );
         }
       }
-      const wrapper = shallow(<Foo id="foo" />);
+      const wrapper = mount(<Foo id="foo" />);
       expect(wrapper.find('.foo').length).to.equal(1);
-      wrapper.setProps({ id: 'bar', foo: 'bla' });
+      wrapper.setProps({ id: 'bar', foo: 'bla'});
       expect(wrapper.find('.bar').length).to.equal(1);
     });
 
@@ -218,8 +222,8 @@ describe('shallow', () => {
           );
         }
       }
-      const nextProps = { id: 'bar', foo: 'bla' };
-      const wrapper = shallow(<Foo id="foo" />);
+      const nextProps = { id: 'bar', foo: 'bla'};
+      const wrapper = mount(<Foo id="foo" />);
       expect(spy.calledOnce).to.be.false;
       wrapper.setProps(nextProps);
       expect(spy.calledOnce).to.be.true;
@@ -242,12 +246,12 @@ describe('shallow', () => {
             <a
               className={`clicks-${this.state.count}`}
               onClick={() => this.setState({ count: this.state.count + 1 })}
-            >foo</a>
+              >foo</a>
           );
         }
       }
 
-      const wrapper = shallow(<Foo />);
+      const wrapper = mount(<Foo />);
 
       expect(wrapper.find('.clicks-0').length).to.equal(1);
       wrapper.simulate('click');
@@ -266,13 +270,11 @@ describe('shallow', () => {
         }
       }
 
-      const wrapper = shallow(<Foo />);
-      const a = {};
-      const b = {};
+      const wrapper = mount(<Foo />);
 
-      wrapper.simulate('click', a, b);
-      expect(spy.args[0][0]).to.equal(a);
-      expect(spy.args[0][1]).to.equal(b);
+      wrapper.simulate('click', { someSpecialData: 'foo'});
+      expect(spy.calledOnce).to.be.true;
+      expect(spy.args[0][0].someSpecialData).to.equal('foo');
     });
 
   });
@@ -290,33 +292,52 @@ describe('shallow', () => {
           );
         }
       }
-      const wrapper = shallow(<Foo />);
+      const wrapper = mount(<Foo />);
       expect(wrapper.find('.foo').length).to.equal(1);
       wrapper.setState({ id: 'bar' });
       expect(wrapper.find('.bar').length).to.equal(1);
+    });
+
+    it('allows setState inside of componentDidMount', () => {
+      // NOTE: this test is a test to ensure that the following issue is
+      // fixed: https://github.com/airbnb/enzyme/issues/27
+      class MySharona extends React.Component {
+        constructor(props) {
+          super(props);
+          this.state = {mounted: false};
+        }
+        componentDidMount() {
+          this.setState({mounted: true});
+        }
+        render() {
+          return <div>{this.state.mounted ? 'a' : 'b'}</div>;
+        }
+      }
+      const wrapper = mount(<MySharona />);
+      expect(wrapper.find('div').text()).to.equal('a');
     });
   });
 
   describe('.is(selector)', () => {
     it('should return true when selector matches current element', () => {
-      const wrapper = shallow(<div className="foo bar baz" />);
+      const wrapper = mount(<div className="foo bar baz" />);
       expect(wrapper.is('.foo')).to.be.true;
     });
 
     it('should allow for compound selectors', () => {
-      const wrapper = shallow(<div className="foo bar baz" />);
+      const wrapper = mount(<div className="foo bar baz" />);
       expect(wrapper.is('.foo.bar')).to.be.true;
     });
 
     it('should return false when selector does not match', () => {
-      const wrapper = shallow(<div className="bar baz" />);
+      const wrapper = mount(<div className="bar baz" />);
       expect(wrapper.is('.foo')).to.be.false;
     });
   });
 
   describe('.not(selector)', () => {
     it('filters to things not matching a selector', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo bar baz" />
           <div className="foo" />
@@ -334,7 +355,7 @@ describe('shallow', () => {
 
   describe('.filter(selector)', () => {
     it('should return a new wrapper of just the nodes that matched the selector', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo bar baz" />
           <div className="foo" />
@@ -354,7 +375,7 @@ describe('shallow', () => {
     });
 
     it('should only look in the current wrappers nodes, not their children', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo">
             <div className="bar" />
@@ -369,7 +390,7 @@ describe('shallow', () => {
 
   describe('.filterWhere(predicate)', () => {
     it('should filter only the nodes of the wrapper', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo bar" />
           <div className="foo baz" />
@@ -387,8 +408,8 @@ describe('shallow', () => {
       expect(baz.hasClass('baz')).to.be.true;
     });
 
-    it('should call the predicate with the wrapped node as the first argument', () => {
-      const wrapper = shallow(
+    it('should call the predicate with the wrapper as the first argument', () => {
+      const wrapper = mount(
         <div>
           <div className="foo bar" />
           <div className="foo baz" />
@@ -401,9 +422,9 @@ describe('shallow', () => {
       const spy = sinon.spy(stub);
       wrapper.find('.foo').filterWhere(spy);
       expect(spy.callCount).to.equal(3);
-      expect(spy.args[0][0]).to.be.instanceOf(ShallowWrapper);
-      expect(spy.args[1][0]).to.be.instanceOf(ShallowWrapper);
-      expect(spy.args[2][0]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[0][0]).to.be.instanceOf(ReactWrapper);
+      expect(spy.args[1][0]).to.be.instanceOf(ReactWrapper);
+      expect(spy.args[2][0]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[0][0].hasClass('bar')).to.be.true;
       expect(spy.args[1][0].hasClass('baz')).to.be.true;
       expect(spy.args[2][0].hasClass('bux')).to.be.true;
@@ -413,13 +434,13 @@ describe('shallow', () => {
   describe('.text()', () => {
 
     const matchesRender = function matchesRender(node) {
-      const actual = shallow(node).text();
+      const actual = mount(node).text();
       const expected = render(node).text();
       expect(expected).to.equal(actual);
     };
 
     it('should handle simple text nodes', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>some text</div>
       );
       expect(wrapper.text()).to.equal('some text');
@@ -445,17 +466,17 @@ describe('shallow', () => {
       );
     });
 
-    it('should render composite components dumbly', () => {
+    it('should render composite components smartly', () => {
       class Foo extends React.Component {
-        render() { return <div/>; }
+        render() { return <div>foo</div>; }
       }
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <Foo />
           <div>test</div>
         </div>
       );
-      expect(wrapper.text()).to.equal('<Foo />test');
+      expect(wrapper.text()).to.equal('footest');
     });
 
     it('should handle html entities', () => {
@@ -468,7 +489,7 @@ describe('shallow', () => {
 
     it('should return the props object', () => {
       const fn = ()=>{};
-      const wrapper = shallow(
+      const wrapper = mount(
         <div id="fooId" className="bax" onClick={fn} >
           <div className="baz" />
           <div className="foo" />
@@ -483,7 +504,7 @@ describe('shallow', () => {
 
     it('should be allowed to be used on an inner node', () => {
       const fn = ()=>{};
-      const wrapper = shallow(
+      const wrapper = mount(
         <div className="bax">
           <div className="baz" onClick={fn} />
           <div className="foo" id="fooId" />
@@ -505,7 +526,7 @@ describe('shallow', () => {
         }
         render() { return <div />; }
       }
-      const wrapper = shallow(<Foo />);
+      const wrapper = mount(<Foo />);
       expect(wrapper.state()).to.eql({ foo: 'foo' });
     });
 
@@ -517,7 +538,7 @@ describe('shallow', () => {
         }
         render() { return <div />; }
       }
-      const wrapper = shallow(<Foo />);
+      const wrapper = mount(<Foo />);
       wrapper.setState({ foo: 'bar' });
       expect(wrapper.state()).to.eql({ foo: 'bar' });
     });
@@ -530,19 +551,19 @@ describe('shallow', () => {
         }
         render() { return <div />; }
       }
-      const wrapper = shallow(<Foo />);
+      const wrapper = mount(<Foo />);
       expect(wrapper.state('foo')).to.equal('foo');
     });
   });
 
   describe('.children([selector])', () => {
     it('should return empty wrapper for node with no children', () => {
-      const wrapper = shallow(<div />);
+      const wrapper = mount(<div />);
       expect(wrapper.children().length).to.equal(0);
     });
 
     it('should return the children nodes of the root', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo" />
           <div className="bar" />
@@ -556,7 +577,7 @@ describe('shallow', () => {
     });
 
     it('should not return any of the children of children', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo">
             <div className="bar" />
@@ -580,7 +601,7 @@ describe('shallow', () => {
           );
         }
       }
-      const wrapper = shallow(
+      const wrapper = mount(
         <Foo items={[
           <i key={1} className="bar">abc</i>,
           <i key={2} className="baz">def</i>,
@@ -593,7 +614,7 @@ describe('shallow', () => {
     });
 
     it('should optionally allow a selector to filter by', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo" />
           <div className="bar bip" />
@@ -609,7 +630,7 @@ describe('shallow', () => {
 
   describe('.parents([selector])', () => {
     it('should return an array of current nodes ancestors', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div className="bax">
           <div className="foo">
             <div className="bar">
@@ -629,7 +650,7 @@ describe('shallow', () => {
     });
 
     it('should work for non-leaf nodes as well', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div className="bax">
           <div className="foo">
             <div className="bar">
@@ -647,7 +668,7 @@ describe('shallow', () => {
     });
 
     it('should optionally allow a selector', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div className="bax foo">
           <div className="foo">
             <div className="bar">
@@ -667,7 +688,7 @@ describe('shallow', () => {
 
   describe('.parent()', () => {
     it('should return only the immediate parent of the node', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div className="bax">
           <div className="foo">
             <div className="bar">
@@ -681,7 +702,7 @@ describe('shallow', () => {
     });
 
     it('should work for multiple nodes', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo">
             <div className="baz" />
@@ -705,7 +726,7 @@ describe('shallow', () => {
 
   describe('.closest(selector)', () => {
     it('should return the closest ancestor for a given selector', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div className="foo">
           <div className="foo baz">
             <div className="bax">
@@ -721,7 +742,7 @@ describe('shallow', () => {
     });
 
     it('should only ever return a wrapper of a single node', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div className="bax">
           <div className="foo">
             <div className="bar">
@@ -735,7 +756,7 @@ describe('shallow', () => {
     });
 
     it('should return itself if matching', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div className="bax">
           <div className="foo">
             <div className="baz">
@@ -751,7 +772,7 @@ describe('shallow', () => {
 
   describe('.hasClass(className)', () => {
     it('should return whether or not node has a certain class', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div className="foo bar baz some-long-string FoOo" />
       );
 
@@ -766,7 +787,7 @@ describe('shallow', () => {
 
   describe('.forEach(fn)', () => {
     it('should call a function for each node in the wrapper', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo bax" />
           <div className="foo bar" />
@@ -778,18 +799,18 @@ describe('shallow', () => {
       wrapper.find('.foo').forEach(spy);
 
       expect(spy.callCount).to.equal(3);
-      expect(spy.args[0][0]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[0][0]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[0][0].hasClass('bax')).to.be.true;
-      expect(spy.args[1][0]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[1][0]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[1][0].hasClass('bar')).to.be.true;
-      expect(spy.args[2][0]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[2][0]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[2][0].hasClass('baz')).to.be.true;
     });
   });
 
   describe('.map(fn)', () => {
     it('should call a function with a wrapper for each node in the wrapper', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo bax" />
           <div className="foo bar" />
@@ -801,16 +822,16 @@ describe('shallow', () => {
       wrapper.find('.foo').map(spy);
 
       expect(spy.callCount).to.equal(3);
-      expect(spy.args[0][0]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[0][0]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[0][0].hasClass('bax')).to.be.true;
-      expect(spy.args[1][0]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[1][0]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[1][0].hasClass('bar')).to.be.true;
-      expect(spy.args[2][0]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[2][0]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[2][0].hasClass('baz')).to.be.true;
     });
 
     it('should return an array with the mapped values', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo bax" />
           <div className="foo bar" />
@@ -829,7 +850,7 @@ describe('shallow', () => {
 
   describe('.reduce(fn[, initialValue])', () => {
     it('should call a function with a wrapper for each node in the wrapper', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo bax" />
           <div className="foo bar" />
@@ -841,16 +862,16 @@ describe('shallow', () => {
       wrapper.find('.foo').reduce(spy, 0);
 
       expect(spy.callCount).to.equal(3);
-      expect(spy.args[0][1]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[0][1]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[0][1].hasClass('bax')).to.be.true;
-      expect(spy.args[1][1]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[1][1]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[1][1].hasClass('bar')).to.be.true;
-      expect(spy.args[2][1]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[2][1]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[2][1].hasClass('baz')).to.be.true;
     });
 
     it('should accumulate a value', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div id="bax" className="foo qoo" />
           <div id="bar" className="foo boo" />
@@ -875,7 +896,7 @@ describe('shallow', () => {
 
   describe('.reduceRight(fn[, initialValue])', () => {
     it('should call a function with a wrapper for each node in the wrapper in reverse', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo bax" />
           <div className="foo bar" />
@@ -887,16 +908,16 @@ describe('shallow', () => {
       wrapper.find('.foo').reduceRight(spy, 0);
 
       expect(spy.callCount).to.equal(3);
-      expect(spy.args[0][1]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[0][1]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[0][1].hasClass('baz')).to.be.true;
-      expect(spy.args[1][1]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[1][1]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[1][1].hasClass('bar')).to.be.true;
-      expect(spy.args[2][1]).to.be.instanceOf(ShallowWrapper);
+      expect(spy.args[2][1]).to.be.instanceOf(ReactWrapper);
       expect(spy.args[2][1].hasClass('bax')).to.be.true;
     });
 
     it('should accumulate a value', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div id="bax" className="foo qoo" />
           <div id="bar" className="foo boo" />
@@ -921,7 +942,7 @@ describe('shallow', () => {
 
   describe('.some(selector)', () => {
     it('should return if a node matches a selector', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo qoo" />
           <div className="foo boo" />
@@ -936,7 +957,7 @@ describe('shallow', () => {
 
   describe('.someWhere(predicate)', () => {
     it('should return if a node matches a predicate', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo qoo" />
           <div className="foo boo" />
@@ -951,7 +972,7 @@ describe('shallow', () => {
 
   describe('.every(selector)', () => {
     it('should return if every node matches a selector', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo qoo" />
           <div className="foo boo" />
@@ -966,7 +987,7 @@ describe('shallow', () => {
 
   describe('.everyWhere(predicate)', () => {
     it('should return if every node matches a predicate', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo qoo" />
           <div className="foo boo" />
@@ -981,7 +1002,7 @@ describe('shallow', () => {
 
   describe('.flatMap(fn)', () => {
     it('should return a wrapper with the mapped and flattened nodes', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="foo">
             <div className="bar" />
@@ -1010,38 +1031,9 @@ describe('shallow', () => {
     });
   });
 
-  describe('.shallow()', () => {
-
-    it('should return a shallow rendered instance of the current node', () => {
-      class Bar extends React.Component {
-        render() {
-          return (
-            <div>
-              <div className="in-bar" />
-            </div>
-          );
-        }
-      }
-      class Foo extends React.Component {
-        render() {
-          return (
-            <div>
-              <Bar />
-            </div>
-          );
-        }
-      }
-      const wrapper = shallow(<Foo />);
-      expect(wrapper.find('.in-bar')).to.have.length(0);
-      expect(wrapper.find(Bar)).to.have.length(1);
-      expect(wrapper.find(Bar).shallow().find('.in-bar')).to.have.length(1);
-    });
-
-  });
-
   describe('.first()', () => {
     it('should return the first node in the current set', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="bar baz" />
           <div className="bar" />
@@ -1055,7 +1047,7 @@ describe('shallow', () => {
 
   describe('.last()', () => {
     it('should return the last node in the current set', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="bar" />
           <div className="bar" />
@@ -1069,7 +1061,7 @@ describe('shallow', () => {
 
   describe('.isEmpty()', () => {
     it('should return true iff wrapper is empty', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div className="foo" />
       );
       expect(wrapper.find('.bar').isEmpty()).to.be.true;
@@ -1079,7 +1071,7 @@ describe('shallow', () => {
 
   describe('.at(index)', () => {
     it('gets a wrapper of the node at the specified index', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="bar foo" />
           <div className="bar bax" />
@@ -1096,7 +1088,7 @@ describe('shallow', () => {
 
   describe('.get(index)', () => {
     it('gets the node at the specified index', () => {
-      const wrapper = shallow(
+      const wrapper = mount(
         <div>
           <div className="bar foo" />
           <div className="bar bax" />
@@ -1109,76 +1101,6 @@ describe('shallow', () => {
       expect(wrapper.find('.bar').get(2)).to.equal(wrapper.find('.bux').node);
       expect(wrapper.find('.bar').get(3)).to.equal(wrapper.find('.baz').node);
     });
-  });
-
-  describe('.debug()', () => {
-    it('should pass through to the debugNodes function', () => {
-      expect(shallow(<div />).debug()).to.equal('<div />');
-    });
-  });
-
-  describe('.html()', () => {
-    it('should return html of straight DOM elements', () => {
-      const wrapper = shallow(
-        <div className="test">
-          <span>Hello World!</span>
-        </div>
-      );
-      expect(wrapper.html()).to.equal(
-        `<div class="test"><span>Hello World!</span></div>`
-      );
-    });
-
-    it('should render out nested composite components', () => {
-      class Foo extends React.Component {
-        render() {
-          return (<div className="in-foo" />);
-        }
-      }
-      class Bar extends React.Component {
-        render() {
-          return (
-            <div className="in-bar">
-              <Foo />
-            </div>
-          );
-        }
-      }
-      const wrapper = shallow(<Bar />);
-      expect(wrapper.html()).to.equal(
-        `<div class="in-bar"><div class="in-foo"></div></div>`
-      );
-      expect(wrapper.find(Foo).html()).to.equal(
-        `<div class="in-foo"></div>`
-      );
-    });
-
-  });
-
-  describe('.render()', () => {
-
-    it('should return a cheerio wrapper around the current node', () => {
-      class Foo extends React.Component {
-        render() {
-          return (<div className="in-foo" />);
-        }
-      }
-      class Bar extends React.Component {
-        render() {
-          return (
-            <div className="in-bar">
-              <Foo />
-            </div>
-          );
-        }
-      }
-      const wrapper = shallow(<Bar />);
-      expect(wrapper.render().find('.in-bar')).to.have.length(1);
-      const renderedFoo = wrapper.find(Foo).render();
-      expect(renderedFoo.find('.in-foo')).to.have.length(1);
-      expect(renderedFoo.find('.in-bar')).to.have.length(0);
-    });
-
   });
 
 });
